requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
django-cors-headers==3.13.0

